agent/requirements.txt
python-socketio[client]
requests
psutil
